Mitigation was added to Section 3.3.2 - Security Assurance to ensure completeness
Mitigation was added to Section 3.3.2 - Security Assurance to ensure completeness, as remediation was mentioned in this section previously.

We previously talked said "document necessary remediation steps suitable for the use-case of the software" but did not talk about the alternative action of mitigation. To ensure completeness, this was added.

Discuss North America / Asia call 2023-02-21

See https://github.com/OpenChain-Project/Security-Assurance-Specification/issues/26