Update openchain-security-specification-2.0.md
On the call 2023-03-21 we revisited the first part of SZ's suggestion for changing respond to receive, and concluded that the language needed improvement to deal with both, but with making response optional. The conclusion is that we will use the following language (pending overall spec improvements to language clarity and compactness):

3.2.1 - Access
Maintain a process to effectively respond to Known Vulnerability external inquiries. Publicly identify a means by which a third party can inquire about a Known Vulnerability with respect to a given software offering.

3.2.1 - Access
Maintain a process to effectively receive and if appropriate respond to Known Vulnerability external inquiries. Publicly identify a means by which a third party can inquire about a Known Vulnerability with respect to a given software offering.

https://github.com/OpenChain-Project/Security-Assurance-Specification/issues/19