Update openchain-security-specification-2.0.md
Feedback discussed to adjust language and resolve this issue:
https://github.com/OpenChain-Project/Security-Assurance-Specification/issues/12

Section 3.3.1 Intro changed from:

A process shall exist for creating and maintaining a bill of materials that includes each Open Source Software component from which the Supplied Software is comprised.

To:

A process shall exist for creating and maintaining a bill of materials that includes each Open Source Software component incorporated in the Supplied Software.

Section 3.3.2 Rationale changed from:

To ensure the Program is sufficiently robust to handle the identified Known Vulnerabilities for the Open Source Software from which the Supplied Software is comprised. That a procedure exists to support this activity and that the procedure is followed.

To ensure the Program is sufficiently robust to handle the identified Known Vulnerabilities for the Open Source Software included in the Supplied Software. That a procedure exists to support this activity and that the procedure is followed.