Update openchain-security-specification-2.0.md
### 3.3.2 - Security Assurance

A process shall exist to detect, identify, and document the existence of Known Vulnerabilities in each Open Source Software component on the Software Bill of Materials (SBOM) for the Supplied Software.

To

### 3.3.2 - Security Assurance

A process shall exist to detect, identify, and document the existence of Known Vulnerabilities in each Open Source Software component in the Software Bill of Materials (SBOM) for the Supplied Software.

Signed-off-by: Shane Coughlan mailto:scoughlan@linuxfoundation.org